Add files via upload
Hey so this code scrapes the text from a user 250 most recent tweets and it then generates a barplot of the top 10. It's a slow program, so I'm hoping you guys can optimize it. Also, the massive array I have in one of the methods is because I couldn't import the 'nltk' library and use the scrapewords function.